--- src/main/java/com/domain/orders/entity/OrderStatus.java
@@ -1,6 +1,0 @@
-package com.domain.orders.entity;
-
-public enum OrderStatus {
-    READY,
-    COMPLETED
-}
