+++ src/main/java/domain/orders/controller/OrderController.java
@@ -1,0 +1,9 @@
+package domain.orders.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+
+@Controller
+@RequiredArgsConstructor
+public class OrderController {
+}
