+++ src/main/java/domain/orders/dto/OrderCreateRequest.java
@@ -1,0 +1,4 @@
+package domain.orders.dto;
+
+public class OrderCreateRequest {
+}
