+++ src/main/java/domain/orders/dto/OrderResponse.java
@@ -1,0 +1,4 @@
+package domain.orders.dto;
+
+public class OrderResponse {
+}
