+++ src/main/java/domain/orders/dto/OrderStatusUpdateRequest.java
@@ -1,0 +1,4 @@
+package domain.orders.dto;
+
+public class OrderStatusUpdateRequest {
+}
