+++ src/main/java/domain/orders/entity/OrderStatus.java
@@ -1,0 +1,6 @@
+package domain.orders.entity;
+
+public enum OrderStatus {
+    READY,
+    COMPLETED
+}
