+++ src/main/java/domain/orders/repository/OrderRepository.java
@@ -1,0 +1,8 @@
+package domain.orders.repository;
+
+
+import domain.orders.entity.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
