+++ src/main/java/domain/orders/service/OrderService.java
@@ -1,0 +1,9 @@
+package domain.orders.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class OrderService {
+}
